+++ src/codeactions/addFields/alAddXmlPortFieldsCodeCommand.ts
@@ -18,7 +18,7 @@ export class ALAddXmlPortFieldsCodeCommand extends ALBaseAddFieldsCodeCommand {
         this.commandTitle = newCommandTitle;
     }
 
-    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation | undefined, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[]) {
+    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation | undefined, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[], onSaveEdit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {
         if ((symbol) &&         
             ((symbol.kind == AZSymbolKind.XmlPortTableElement) ||
              (symbol.kind == AZSymbolKind.XmlPortFieldElement) ||
@@ -31,6 +31,8 @@ export class ALAddXmlPortFieldsCodeCommand extends ALBaseAddFieldsCodeCommand {
             };
             actions.push(action);
         }
+
+        return onSaveEdit;
     }
 
     protected async runAsync(docSymbols: AZDocumentSymbolsLibrary, document: vscode.TextDocument, range: vscode.Range) {