+++ src/codeactions/sortSymbols/alSortProceduresCodeCommand.ts
@@ -32,31 +32,31 @@ export class ALSortProceduresCodeCommand extends ALBaseSortCodeCommand {
         super(context, "SortProcedures");
     }
 
-    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation | undefined, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[]) {
-        let edit: vscode.WorkspaceEdit | undefined = undefined;
-        let fixOnSave = this.canRunOnSave(document.uri);
-        let actionKind = this.getCodeActionKind(fixOnSave);
-
-        if (fixOnSave) {
+    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation | undefined, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[], onSaveEdit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {
+        if (this.canRunOnSave(document.uri)) {
             let objList: AZSymbolInformation[] = [];
             if (docSymbols.rootSymbol) {
                 docSymbols.rootSymbol.collectObjectSymbols(objList);
                 for (let i=0; i<objList.length; i++)
-                    edit = this.prepareEdit(objList[i], document, edit);
+                    onSaveEdit = this.prepareEdit(objList[i], document, onSaveEdit);
             }
         } else {
+            let edit: vscode.WorkspaceEdit | undefined = undefined;
+
             if ((symbol) &&
                 ((this.isMethodSymbol(symbol)) || (symbol.isALObject())) &&
                 (symbol.selectionRange) &&
                 (symbol.selectionRange.start.line == range.start.line))
                 edit = this.prepareEdit(symbol, document, edit);
+
+            if (edit) {
+                let action = new vscode.CodeAction("Sort procedures", vscode.CodeActionKind.QuickFix);
+                action.edit = edit;
+                actions.push(action);
+            }        
         }
 
-        if (edit) {
-            let action = new vscode.CodeAction("Sort procedures", actionKind);
-            action.edit = edit;
-            actions.push(action);
-        }
+        return onSaveEdit;
     }
 
     protected prepareEdit(symbol: AZSymbolInformation, document: vscode.TextDocument, edit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {